Move CLI path helpers into shared utils for reuse
Moved the generic helpers from src/cli/lib/path-utils.js into src/shared/utils/path.js so path containment checks live alongside the shared path utilities and can be reused outside the CLI.
